+++ app/Http/Livewire/ScholarshipRequirementEditItemOptionLivewire.php
@@ -4,6 +4,7 @@ namespace App\Http\Livewire;
 
 use Livewire\Component;
 use App\Models\ScholarshipRequirementItemOption;
+use Illuminate\Support\Facades\Auth;
 
 class ScholarshipRequirementEditItemOptionLivewire extends Component
 {
@@ -16,14 +17,28 @@ class ScholarshipRequirementEditItemOptionLivewire extends Component
         'option.option' => 'required|string|min:1|max:240',
     ];
 
+    protected function verifyUser()
+    {
+        if (!Auth::check()) {
+            redirect()->route('index');
+            return true;
+        }
+        return false;
+    }
+    
+
     public function mount(ScholarshipRequirementItemOption $id, $type)
     {
+        if ($this->verifyUser()) return;
+
         $this->option = $id;
         $this->type = $type;
     }
     
     public function updated($propertyName)
     {
+        if ($this->verifyUser()) return;
+
         $this->validate();
         $this->option->save();
     }
@@ -35,6 +50,8 @@ class ScholarshipRequirementEditItemOptionLivewire extends Component
 
     public function delete_option()
     {
+        if ($this->verifyUser()) return;
+
         $this->option->delete();
         $this->dispatchBrowserEvent('delete_option_div', [
             'div_class' => $this->option->id,  
@@ -44,6 +61,8 @@ class ScholarshipRequirementEditItemOptionLivewire extends Component
 
     public function save_all()
     {
+        if ($this->verifyUser()) return;
+        
         $this->validate();
         $this->option->save();
     }
